Add more docstrings
