+++ tests/test_manipulate_files.py
@@ -19,6 +19,7 @@ class FiletoteManipulateFiles(FiletoteTestCase):
     """
 
     def setUp(self):
+        """Provides shared setup for tests."""
         super().setUp()
 
         self._create_flat_import_dir()
