+++ tests/test_pruning.py
@@ -19,6 +19,7 @@ class FiletotePruningyTest(FiletoteTestCase):
     """
 
     def setUp(self):
+        """Provides shared setup for tests."""
         super().setUp()
 
         self._create_flat_import_dir()
